requirements.txt
accelerate @ git+https://github.com/huggingface/accelerate.git@a2d8f540c3ab37c8f84d616be1300a0572b69cf8
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
bitsandbytes==0.40.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
chainlit==0.7.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.25.0
dataclasses-json==0.5.14
datasets==2.13.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.6
docx2txt==0.8
emoji==2.8.0
exceptiongroup==1.1.3
fastapi==0.100.1
fastapi-socketio==0.0.10
filelock==3.9.0
filetype==1.2.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.309
langdetect==1.0.9
langsmith==0.0.43
Lazify==0.4.0
lit==15.0.7
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
nodeenv==1.8.0
numpy==1.25.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.1
pandas==2.0.3
peft @ git+https://github.com/huggingface/peft.git@e06d94ddeb6c70913593740618df76908b918d66
pgvector==0.2.3
Pillow==9.5.0
prisma==0.10.0
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.9
pyarrow==12.0.1
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pypdf==3.16.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.7.1
python-graphql-client==0.4.3
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.1
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
streamlit==1.25.0
sympy==1.11.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.19
uptrace==1.20.0
urllib3==2.0.4
uvicorn==0.23.2
validators==0.20.0
watchdog==3.0.0
watchfiles==0.20.0
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.15.0
wsproto==1.2.0
xformers==0.0.22
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
